requirements.txt
autopep8==1.1.1
awesome-slugify==1.6.4
cffi==1.0.3
coverage==3.7.1
cryptography==0.9
defusedxml==0.4.1
diff-match-patch==20121119
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.2
django-autoslug==1.8.0
django-cities-light==3.1.2
django-colorful==1.0.1
django-datetime-widget==0.9.3
django-debug-toolbar==1.3.0
django-extensions==1.5.5
django-filter==0.10.0
django-import-export==0.2.7
django-redis-cache==0.13.1
django-rest-swagger==0.2.9
django-silk==0.5.2
django-toolbelt==0.0.1
djangorestframework==3.1.2
enum34==1.0.4
gunicorn==19.3.0
idna==2.0
Jinja2==2.7.3
Markdown==2.6.2
MarkupSafe==0.23
model-mommy==1.2.3
ndg-httpsclient==0.4.0
oauthlib==0.7.2
pep8==1.6.2
psycopg2==2.6
pusher==1.2.0
pyasn1==0.1.7
pycparser==2.13
Pygments==2.0.2
PyJWT==1.3.0
pyOpenSSL==0.15.1
pystache==0.5.4
python-dateutil==2.4.2
python-openid==2.2.5
python-social-auth==0.1.22
python3-openid==3.0.5
pytz==2015.4
PyYAML==3.11
redis==2.10.3
regex==2015.5.28
requests==2.7.0
requests-oauthlib==0.4.2
selenium==2.45.0
simplejson==3.7.2
six==1.9.0
sleekxmpp==1.3.1
sqlparse==0.1.15
static==1.1.1
static3==0.6.1
tablib==0.10.0
Unidecode==0.4.17
urllib3==1.10.4
